Add assert statements
The main bulk of the app only fails because of improper user input.

But minor details such as improper input format in storage
and non existent task list are potential programmer bugs which need to be
handled.

Let's ensure such assumptions are handled properly by adding assert
statements in the corresponding methods.