Allow users to find free times
Current implementation does not allow users to find free times
to help them better plan their schedule.

Let's create a free command which allows users to input either a
date or a time period and find out whether their schedule is clear.

If they are busy, the chatbot will tell them the tasks which they are
busy with.

The free command takes in a LocalDate instead of a LocalDateTime,
so a bug is when a task's end time is 00:00, and so the user is
technically free on that date, but the program states that they are
busy.